+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.5.11
 
 - Adding `JsonbListView` with `isSqlNull(int index)` method to check if a JSONB array has SQL or JSON null value.
+- Fix double-timezone encoding bug.
 
 ## 3.5.10